# Conflicts:
#	db/schema.rb
#	spec/system/admin/organizations_system_spec.rb
#	spec/system/request_system_spec.rb

+++ app/services/transfer_destroy_service.rb
@@ -4,8 +4,13 @@ class TransferDestroyService
   end
 
   def call
+    if Audit.finalized_since?(transfer, transfer.to_id, transfer.from_id)
+      raise "We can't delete this transfer because its items were audited since you made the transfer."
+    end
+
     transfer.transaction do
       revert_inventory_transfer!
+      TransferDestroyEvent.publish(transfer)
       transfer.destroy!
     end
 
@@ -23,7 +28,7 @@ class TransferDestroyService
   end
 
   def revert_inventory_transfer!
-    transfer.to.decrease_inventory(transfer)
-    transfer.from.increase_inventory(transfer)
+    transfer.to.decrease_inventory(transfer.line_item_values)
+    transfer.from.increase_inventory(transfer.line_item_values)
   end
 end
